Add experimental addition for serial prefix summing in radixsort
This is done when there are more items to prefix sum than threads. It
seems to make a big improvement on GCN, but it is also preventing loop
unrolling from happening in some cases. This is work being parked here
in preparation for having the host generate the upsweep/downsweep
function call sequence.