Experimental code for making the scan block count a multiple of compute units.
This seems to go badly: the final performance for 50e6 elements is at
least 10% worse on Pitcairn than when using a power-of-two block count,
even with both choices thrown into the autotuner. Some more
investigation is required: possibly the problem sizes chosen for tuning
are poor, or maybe 50e6 is a freak case. Also, it might need to be the
right multiple of the compute units, to exactly fill up the occupancy on
each CU.